ALL COMMANDS STORED INDIVIDUALLY
