+++ commands/ping.js
@@ -1,10 +1,10 @@
 const { SlashCommandBuilder } = require('discord.js');
 
 module.exports = {
-	data: new SlashCommandBuilder()
-		.setName('ping')
-		.setDescription('Replies with Pong!'),
-	async execute(interaction) {
-		await interaction.reply('Pong!');
-	},
+    data: new SlashCommandBuilder()
+        .setName('ping') // Name must be lowercase
+        .setDescription('Replies with Pong!'), // Description is required
+    async execute(interaction) {
+        await interaction.reply('Pong?');
+    },
 };
